requirements.txt
Flask
Flask-Mail
psycopg2
SQLAlchemy
python-dotenv
gunicorn